phase-5/services/recurring-task-service/requirements.txt
confluent-kafka==2.6.0
sqlmodel==0.0.22
pydantic==2.10.1
python-dateutil==2.9.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10